moving action functions to different class, and to parsing class, IN DEVELOPMENT
